refactor(containers): split container view model by responsibility
