--- ArcBox/ViewModels/ContainersViewModel.swift
@@ -1,855 +1,0 @@
-import ArcBoxClient
-import DockerClient
-import OSLog
-import SwiftUI
-
-extension Notification.Name {
-    /// Posted when Docker resources change (e.g. container deleted) so other sections can refresh.
-    static let dockerDataChanged = Notification.Name("dockerDataChanged")
-}
-
-/// Detail panel tab for containers
-enum ContainerDetailTab: String, CaseIterable, Identifiable {
-    case info = "Info"
-    case logs = "Logs"
-    case terminal = "Terminal"
-    case files = "Files"
-
-    var id: String { rawValue }
-}
-
-/// Sort field for containers
-enum ContainerSortField: String, CaseIterable {
-    case name = "Name"
-    case dateCreated = "Date Created"
-    case status = "Status"
-}
-
-/// Container list loading state
-enum ContainerLoadState: Equatable {
-    case waiting  // Waiting for docker client
-    case loading  // Fetching from Docker API
-    case loaded  // Fetch completed (containers may be empty)
-    case failed(String)  // Fetch failed with error message
-}
-
-struct ContainerCreateOptions {
-    let image: String
-    let name: String
-    let platform: String?
-    let command: String
-    let entrypoint: String
-    let workingDir: String
-    let autoRemove: Bool
-    let restartPolicy: String
-    let privileged: Bool
-    let readOnlyRootfs: Bool
-    let dockerInit: Bool
-}
-
-/// Container list state, selection, tabs, grouping
-@MainActor
-@Observable
-class ContainersViewModel {
-    private struct ContainerDetailSnapshot {
-        let domain: String?
-        let ipAddress: String?
-        let mounts: [ContainerMount]
-        let rootfsMountPath: String?
-    }
-
-    var containers: [ContainerViewModel] = []
-    var loadState: ContainerLoadState = .waiting
-    var selectedID: String?
-    var activeTab: ContainerDetailTab = .info
-    var expandedGroups: Set<String> = []
-    var listWidth: CGFloat = 320
-    var searchText: String = ""
-    var isSearching: Bool = false
-    var showNewContainerSheet: Bool = false
-    var sortBy: ContainerSortField = .name
-    var sortAscending: Bool = true
-    var lastError: String?
-
-    var runningCount: Int {
-        containers.filter(\.isRunning).count
-    }
-
-    var selectedContainer: ContainerViewModel? {
-        guard let id = selectedID else { return nil }
-        guard var container = containers.first(where: { $0.id == id }) else { return nil }
-        if let details = detailsByID[id] {
-            container.domain = details.domain
-            container.ipAddress = details.ipAddress
-            container.mounts = details.mounts
-            container.rootfsMountPath = details.rootfsMountPath
-        }
-        return container
-    }
-
-    private var detailsByID: [String: ContainerDetailSnapshot] = [:]
-    private var iconsByImage: [String: String] = [:]
-
-    private func sortedContainers(_ list: [ContainerViewModel]) -> [ContainerViewModel] {
-        list.sorted { a, b in
-            let result: Bool
-            switch sortBy {
-            case .name:
-                result = a.name.localizedCaseInsensitiveCompare(b.name) == .orderedAscending
-            case .dateCreated:
-                result = a.createdAt < b.createdAt
-            case .status:
-                result = a.state.rawValue < b.state.rawValue
-            }
-            return sortAscending ? result : !result
-        }
-    }
-
-    private func matchesSearch(_ container: ContainerViewModel) -> Bool {
-        guard !searchText.isEmpty else { return true }
-        let query = searchText.lowercased()
-        return container.name.lowercased().contains(query)
-            || container.image.lowercased().contains(query)
-            || (container.composeProject?.lowercased().contains(query) ?? false)
-    }
-
-    /// Group containers by compose project
-    var composeGroups: [(project: String, containers: [ContainerViewModel])] {
-        var groups: [String: [ContainerViewModel]] = [:]
-        for container in containers where matchesSearch(container) {
-            if let project = container.composeProject {
-                groups[project, default: []].append(container)
-            }
-        }
-        return groups.sorted { $0.key < $1.key }.map {
-            (project: $0.key, containers: sortedContainers($0.value))
-        }
-    }
-
-    /// Containers without a compose project
-    var standaloneContainers: [ContainerViewModel] {
-        sortedContainers(containers.filter { $0.composeProject == nil && matchesSearch($0) })
-    }
-
-    func selectContainer(_ id: String) {
-        selectedID = id
-    }
-
-    func selectContainer(_ id: String, docker: DockerClient?) async {
-        selectedID = id
-        await loadContainerDetailsFromDocker(id, docker: docker)
-    }
-
-    func selectContainer(_ id: String, client: ArcBoxClient?, docker: DockerClient?) async {
-        selectedID = id
-        if docker != nil {
-            await loadContainerDetailsFromDocker(id, docker: docker)
-        } else {
-            await loadContainerDetails(id, client: client)
-        }
-    }
-
-    func toggleGroup(_ group: String) {
-        if expandedGroups.contains(group) {
-            expandedGroups.remove(group)
-        } else {
-            expandedGroups.insert(group)
-        }
-    }
-
-    func isGroupExpanded(_ group: String) -> Bool {
-        expandedGroups.contains(group)
-    }
-
-    private func applyExpandedGroups(from list: [ContainerViewModel]) {
-        for container in list {
-            if let project = container.composeProject {
-                expandedGroups.insert(project)
-            }
-        }
-    }
-
-    private func setContainerRunningState(_ id: String, isRunning: Bool) {
-        updateContainer(id) { container in
-            container.state = isRunning ? .running : .stopped
-        }
-    }
-
-    private func setTransitioning(_ id: String, _ value: Bool) {
-        updateContainer(id) { container in
-            container.isTransitioning = value
-        }
-    }
-
-    /// IDs currently transitioning, used to preserve state across container reloads
-    private var transitioningIDs: Set<String> {
-        Set(containers.filter(\.isTransitioning).map(\.id))
-    }
-
-    private func removeContainerLocally(_ id: String) {
-        containers.removeAll { $0.id == id }
-        detailsByID.removeValue(forKey: id)
-        if selectedID == id {
-            selectedID = nil
-        }
-    }
-
-    private func setContainerDetails(
-        _ id: String,
-        domain: String?,
-        ipAddress: String?,
-        mounts: [ContainerMount],
-        rootfsMountPath: String? = nil
-    ) {
-        let currentContainer = containers.first(where: { $0.id == id })
-        let labels = currentContainer?.labels ?? [:]
-        let inferredRootfsMountPath = ContainerViewModel.inferRootFSMountPath(
-            explicitPath: rootfsMountPath ?? currentContainer?.rootfsMountPath,
-            labels: labels,
-            mounts: mounts
-        )
-
-        detailsByID[id] = ContainerDetailSnapshot(
-            domain: domain,
-            ipAddress: ipAddress,
-            mounts: mounts,
-            rootfsMountPath: inferredRootfsMountPath
-        )
-        updateContainer(id) { container in
-            container.domain = domain
-            container.ipAddress = ipAddress
-            container.mounts = mounts
-            container.rootfsMountPath = inferredRootfsMountPath
-        }
-    }
-
-    private func updateContainer(_ id: String, mutate: (inout ContainerViewModel) -> Void) {
-        guard let index = containers.firstIndex(where: { $0.id == id }) else { return }
-        var snapshot = containers
-        mutate(&snapshot[index])
-        containers = snapshot
-    }
-
-    private func containerDetailsCache() -> [String: ContainerDetailSnapshot] {
-        Dictionary(
-            uniqueKeysWithValues: detailsByID.map { id, details in
-                (id, details)
-            }
-        )
-    }
-
-    private func applyCachedDetails(
-        _ cache: [String: ContainerDetailSnapshot],
-        to viewModels: inout [ContainerViewModel]
-    ) {
-        for i in viewModels.indices {
-            guard let details = cache[viewModels[i].id] else { continue }
-            viewModels[i].domain = details.domain
-            viewModels[i].ipAddress = details.ipAddress
-            viewModels[i].mounts = details.mounts
-            viewModels[i].rootfsMountPath = details.rootfsMountPath
-        }
-    }
-
-    private func applyCachedIcons(to viewModels: inout [ContainerViewModel]) {
-        for i in viewModels.indices {
-            viewModels[i].iconURL = iconsByImage[viewModels[i].image]
-        }
-    }
-
-    /// Fetch icon URLs for all unique image references that are not already cached.
-    func fetchIcons(client: ArcBoxClient?) async {
-        guard let client else { return }
-        let uncached = Set(containers.map(\.image)).subtracting(iconsByImage.keys)
-        guard !uncached.isEmpty else { return }
-
-        await withTaskGroup(of: (String, String?, Bool).self) { group in
-            for image in uncached {
-                group.addTask {
-                    do {
-                        var request = Arcbox_V1_GetImageIconRequest()
-                        request.fqin = image
-                        let response = try await client.icons.getImageIcon(
-                            request, options: ArcBoxClient.defaultCallOptions)
-                        let url = response.url.isEmpty ? nil : response.url
-                        // (image, url, succeeded) — cache empty url as "no icon available"
-                        return (image, url, true)
-                    } catch {
-                        Log.container.debug(
-                            "Icon fetch failed for \(image, privacy: .private): \(error.localizedDescription, privacy: .private)"
-                        )
-                        // Mark as failed so we don't cache the negative result
-                        return (image, nil, false)
-                    }
-                }
-            }
-            for await (image, url, succeeded) in group {
-                if let url {
-                    iconsByImage[image] = url
-                } else if succeeded {
-                    // RPC succeeded but no icon available — cache to avoid repeated lookups
-                    iconsByImage[image] = ""
-                }
-                // If RPC failed, leave uncached so next load retries
-            }
-        }
-
-        var snapshot = containers
-        applyCachedIcons(to: &snapshot)
-        containers = snapshot
-    }
-
-    // MARK: - gRPC Operations
-
-    /// Load containers from daemon via gRPC.
-    func loadContainers(client: ArcBoxClient?) async {
-        guard let client else {
-            Log.container.debug("No gRPC client available")
-            return
-        }
-
-        let currentTransitioning = transitioningIDs
-        let cachedDetails = containerDetailsCache()
-        do {
-            var request = Arcbox_V1_ListContainersRequest()
-            request.all = true
-            let listRequest = request
-            let response = try await Perf.measure("container.list_grpc") {
-                try await client.containers.list(listRequest, options: ArcBoxClient.defaultCallOptions)
-            }
-            var viewModels = response.containers.map { summary in
-                ContainerViewModel(from: summary)
-            }
-            applyCachedDetails(cachedDetails, to: &viewModels)
-            applyCachedIcons(to: &viewModels)
-            for i in viewModels.indices where currentTransitioning.contains(viewModels[i].id) {
-                viewModels[i].isTransitioning = true
-            }
-            containers = viewModels
-            applyExpandedGroups(from: containers)
-            await fetchIcons(client: client)
-            if let selectedID, containers.contains(where: { $0.id == selectedID }) {
-                await loadContainerDetails(selectedID, client: client)
-            }
-        } catch {
-            Log.container.error("Error loading containers via gRPC: \(error.localizedDescription, privacy: .private)")
-            ErrorReporting.capture(error, domain: .container, operation: "list_grpc")
-        }
-    }
-
-    func startContainer(_ id: String, client: ArcBoxClient?) async {
-        lastError = nil
-        guard let client else { return }
-        setTransitioning(id, true)
-        var request = Arcbox_V1_StartContainerRequest()
-        request.id = id
-        do {
-            _ = try await client.containers.start(request, options: ArcBoxClient.defaultCallOptions)
-            setContainerRunningState(id, isRunning: true)
-        } catch {
-            Log.container.error(
-                "Error starting container \(id, privacy: .private): \(error.localizedDescription, privacy: .private)")
-            ErrorReporting.capture(error, domain: .container, operation: "start")
-            lastError = error.localizedDescription
-        }
-        setTransitioning(id, false)
-        await loadContainers(client: client)
-    }
-
-    func stopContainer(_ id: String, client: ArcBoxClient?) async {
-        lastError = nil
-        guard let client else { return }
-        setTransitioning(id, true)
-        var request = Arcbox_V1_StopContainerRequest()
-        request.id = id
-        do {
-            _ = try await client.containers.stop(request, options: ArcBoxClient.defaultCallOptions)
-            setContainerRunningState(id, isRunning: false)
-        } catch {
-            Log.container.error(
-                "Error stopping container \(id, privacy: .private): \(error.localizedDescription, privacy: .private)")
-            ErrorReporting.capture(error, domain: .container, operation: "stop")
-            lastError = error.localizedDescription
-        }
-        setTransitioning(id, false)
-        await loadContainers(client: client)
-    }
-
-    /// Creates a new container via Docker API and returns its ID, or nil on failure.
-    func createContainer(
-        options: ContainerCreateOptions,
-        docker: DockerClient?
-    ) async -> String? {
-        guard let docker else { return nil }
-        var config = Components.Schemas.ContainerConfig()
-        config.Image = options.image
-        let cmdParts = options.command.split(separator: " ").map(String.init)
-        if !cmdParts.isEmpty { config.Cmd = cmdParts }
-        let entrypointParts = options.entrypoint.split(separator: " ").map(String.init)
-        if !entrypointParts.isEmpty { config.Entrypoint = entrypointParts }
-        if !options.workingDir.isEmpty { config.WorkingDir = options.workingDir }
-
-        let policyName: Components.Schemas.RestartPolicy.NamePayload =
-            switch options.restartPolicy {
-            case "always": .always
-            case "unless-stopped": .unless_hyphen_stopped
-            case "on-failure": .on_hyphen_failure
-            default: .no
-            }
-        var resources = Components.Schemas.Resources()
-        if options.dockerInit { resources.Init = true }
-        let hostConfig = Components.Schemas.HostConfig(
-            value1: resources,
-            value2: .init(
-                RestartPolicy: .init(Name: policyName),
-                AutoRemove: options.autoRemove,
-                Privileged: options.privileged,
-                ReadonlyRootfs: options.readOnlyRootfs
-            )
-        )
-
-        do {
-            let response = try await docker.api.ContainerCreate(
-                query: .init(name: options.name.isEmpty ? nil : options.name, platform: options.platform),
-                body: .json(.init(value1: config, value2: .init(HostConfig: hostConfig)))
-            )
-            switch response {
-            case .created(let created):
-                switch created.body {
-                case .json(let body):
-                    let id = body.Id
-                    Log.container.info("Created container \(id, privacy: .private)")
-                    await loadContainersFromDocker(docker: docker)
-                    return id
-                }
-            case .badRequest(let err):
-                Log.container.error("Bad request creating container: \(String(describing: err), privacy: .private)")
-            case .notFound(let err):
-                Log.container.error("Image not found: \(String(describing: err), privacy: .private)")
-            case .conflict(let err):
-                Log.container.error("Container name conflict: \(String(describing: err), privacy: .private)")
-            case .internalServerError(let err):
-                Log.container.error("Server error creating container: \(String(describing: err), privacy: .private)")
-            case .undocumented(let statusCode, _):
-                Log.container.error("Unexpected status \(statusCode, privacy: .public) creating container")
-            }
-        } catch {
-            Log.container.error("Error creating container: \(String(describing: error), privacy: .private)")
-            ErrorReporting.capture(error, domain: .container, operation: "create")
-        }
-        return nil
-    }
-
-    func removeContainer(_ id: String, client: ArcBoxClient?) async {
-        lastError = nil
-        guard let client else { return }
-        var request = Arcbox_V1_RemoveContainerRequest()
-        request.id = id
-        request.force = true
-        do {
-            _ = try await client.containers.remove(request, options: ArcBoxClient.defaultCallOptions)
-            removeContainerLocally(id)
-        } catch {
-            Log.container.error(
-                "Error removing container \(id, privacy: .private): \(error.localizedDescription, privacy: .private)")
-            ErrorReporting.capture(error, domain: .container, operation: "remove")
-            lastError = error.localizedDescription
-        }
-        await loadContainers(client: client)
-    }
-
-    func loadContainerDetails(_ id: String, client: ArcBoxClient?) async {
-        guard let client else { return }
-
-        var request = Arcbox_V1_InspectContainerRequest()
-        request.id = id
-
-        do {
-            let details = try await client.containers.inspect(request, options: ArcBoxClient.defaultCallOptions)
-            setContainerDetails(
-                id,
-                domain: Self.normalized(details.config.domainname),
-                ipAddress: Self.normalized(details.networkSettings.ipAddress),
-                mounts: details.mounts.map { mount in
-                    ContainerMount(
-                        type: mount.type,
-                        source: mount.source,
-                        destination: mount.destination,
-                        isReadOnly: !mount.rw
-                    )
-                }
-            )
-        } catch {
-            Log.container.error(
-                "Error inspecting container \(id, privacy: .private): \(error.localizedDescription, privacy: .private)")
-            ErrorReporting.capture(error, domain: .container, operation: "inspect_grpc")
-        }
-    }
-
-    // MARK: - Docker API Operations
-
-    /// Load containers from Docker Engine API.
-    func loadContainersFromDocker(docker: DockerClient?, iconClient: ArcBoxClient? = nil) async {
-        guard let docker else {
-            Log.container.debug("No docker client available")
-            return
-        }
-
-        if loadState != .loaded {
-            loadState = .loading
-        }
-
-        let currentTransitioning = transitioningIDs
-        let cachedDetails = containerDetailsCache()
-        do {
-            let containerList = try await Perf.measure("container.list_docker") {
-                let response = try await docker.api.ContainerList(.init(query: .init(all: true)))
-                return try response.ok.body.json
-            }
-            var viewModels = containerList.map { ContainerViewModel(fromDocker: $0) }
-            applyCachedDetails(cachedDetails, to: &viewModels)
-            applyCachedIcons(to: &viewModels)
-            for i in viewModels.indices where currentTransitioning.contains(viewModels[i].id) {
-                viewModels[i].isTransitioning = true
-            }
-            containers = viewModels
-            Log.container.info("Loaded \(self.containers.count, privacy: .public) containers via Docker")
-            applyExpandedGroups(from: containers)
-            await fetchIcons(client: iconClient)
-            if let selectedID, containers.contains(where: { $0.id == selectedID }) {
-                await loadContainerDetailsFromDocker(selectedID, docker: docker)
-            }
-            loadState = .loaded
-        } catch {
-            Log.container.error("Error loading containers: \(error.localizedDescription, privacy: .private)")
-            ErrorReporting.capture(error, domain: .container, operation: "list_docker")
-            if containers.isEmpty {
-                loadState = .failed(error.localizedDescription)
-            } else {
-                loadState = .loaded
-                lastError = error.localizedDescription
-            }
-        }
-    }
-
-    func startContainerDocker(_ id: String, docker: DockerClient?) async {
-        lastError = nil
-        guard let docker else { return }
-        setTransitioning(id, true)
-        do {
-            _ = try await docker.api.ContainerStart(path: .init(id: id))
-            setContainerRunningState(id, isRunning: true)
-        } catch {
-            Log.container.error(
-                "Error starting container \(id, privacy: .private): \(error.localizedDescription, privacy: .private)")
-            ErrorReporting.capture(error, domain: .container, operation: "start_docker")
-            lastError = error.localizedDescription
-        }
-        setTransitioning(id, false)
-        await loadContainersFromDocker(docker: docker)
-    }
-
-    func stopContainerDocker(_ id: String, docker: DockerClient?) async {
-        lastError = nil
-        guard let docker else { return }
-        setTransitioning(id, true)
-        do {
-            _ = try await docker.api.ContainerStop(path: .init(id: id))
-            setContainerRunningState(id, isRunning: false)
-        } catch {
-            Log.container.error(
-                "Error stopping container \(id, privacy: .private): \(error.localizedDescription, privacy: .private)")
-            ErrorReporting.capture(error, domain: .container, operation: "stop_docker")
-            lastError = error.localizedDescription
-        }
-        setTransitioning(id, false)
-        await loadContainersFromDocker(docker: docker)
-    }
-
-    func removeContainerDocker(_ id: String, docker: DockerClient?) async {
-        lastError = nil
-        guard let docker else { return }
-        do {
-            _ = try await docker.api.ContainerDelete(path: .init(id: id), query: .init(force: true))
-            removeContainerLocally(id)
-            NotificationCenter.default.post(name: .dockerDataChanged, object: nil)
-        } catch {
-            Log.container.error(
-                "Error removing container \(id, privacy: .private): \(error.localizedDescription, privacy: .private)")
-            ErrorReporting.capture(error, domain: .container, operation: "remove_docker")
-            lastError = error.localizedDescription
-        }
-        await loadContainersFromDocker(docker: docker)
-    }
-
-    func loadContainerDetailsFromDocker(_ id: String, docker: DockerClient?) async {
-        guard let docker else { return }
-
-        do {
-            // Prefer raw snapshot to avoid date decoding failures and to support
-            // NetworkSettings.Networks.*.IPAddress fallback consistently.
-            let snapshot = try await docker.inspectContainerSnapshot(id: id)
-            let mounts = snapshot.mounts.compactMap { mount -> ContainerMount? in
-                guard let destination = Self.normalized(mount.destination) else { return nil }
-                let source = Self.normalized(mount.source) ?? "-"
-                return ContainerMount(
-                    type: Self.normalized(mount.type) ?? "unknown",
-                    source: source,
-                    destination: destination,
-                    isReadOnly: !(mount.rw ?? true)
-                )
-            }
-            setContainerDetails(
-                id,
-                domain: Self.normalized(snapshot.domainname),
-                ipAddress: Self.normalized(snapshot.ipAddress),
-                mounts: mounts,
-                rootfsMountPath: Self.normalized(snapshot.rootfsMountPath)
-            )
-            let snapshotDomain = Self.normalized(snapshot.domainname) ?? "-"
-            let snapshotIP = Self.normalized(snapshot.ipAddress) ?? "-"
-            let snapshotRootFS = Self.normalized(snapshot.rootfsMountPath) ?? "-"
-            Log.container.debug("Inspect snapshot for \(id, privacy: .private)")
-            Log.container.debug(
-                "domain=\(snapshotDomain, privacy: .private), ip=\(snapshotIP, privacy: .private), mounts=\(mounts.count, privacy: .public)"
-            )
-            Log.container.debug(
-                "rootfs=\(snapshotRootFS, privacy: .private)"
-            )
-        } catch {
-            Log.container.error(
-                "Inspect snapshot failed for \(id, privacy: .private): \(error.localizedDescription, privacy: .private)"
-            )
-            do {
-                // Fallback to generated inspect model if raw path fails unexpectedly.
-                let response = try await docker.api.ContainerInspect(path: .init(id: id))
-                let details = try response.ok.body.json
-
-                let mounts = (details.Mounts ?? []).compactMap { mount -> ContainerMount? in
-                    guard let destination = Self.normalized(mount.Destination) else { return nil }
-                    let source = Self.normalized(mount.Source) ?? "-"
-                    return ContainerMount(
-                        type: "unknown",
-                        source: source,
-                        destination: destination,
-                        isReadOnly: false
-                    )
-                }
-
-                setContainerDetails(
-                    id,
-                    domain: Self.normalized(details.Config?.Domainname),
-                    ipAddress: Self.normalized(details.NetworkSettings?.IPAddress),
-                    mounts: mounts
-                )
-                let fallbackDomain = Self.normalized(details.Config?.Domainname) ?? "-"
-                let fallbackIP = Self.normalized(details.NetworkSettings?.IPAddress) ?? "-"
-                Log.container.debug("Inspect fallback for \(id, privacy: .private)")
-                Log.container.debug(
-                    "domain=\(fallbackDomain, privacy: .private), ip=\(fallbackIP, privacy: .private), mounts=\(mounts.count, privacy: .public)"
-                )
-            } catch {
-                Log.container.error(
-                    "Inspect fallback failed for \(id, privacy: .private): \(error.localizedDescription, privacy: .private)"
-                )
-                ErrorReporting.capture(error, domain: .container, operation: "inspect_docker")
-            }
-        }
-    }
-
-    // MARK: - Batch Docker Operations
-
-    func startContainersDocker(_ ids: [String], docker: DockerClient?) async {
-        guard let docker else { return }
-        let stoppedIDs = ids.filter { id in
-            containers.first(where: { $0.id == id })?.isRunning == false
-        }
-        for id in stoppedIDs { setTransitioning(id, true) }
-        await withTaskGroup(of: Void.self) { group in
-            for id in stoppedIDs {
-                group.addTask { [weak self] in
-                    do {
-                        _ = try await docker.api.ContainerStart(path: .init(id: id))
-                        await self?.setContainerRunningState(id, isRunning: true)
-                    } catch {
-                        Log.container.error(
-                            "Error starting container \(id, privacy: .private): \(error.localizedDescription, privacy: .private)"
-                        )
-                        ErrorReporting.capture(error, domain: .container, operation: "batch_start")
-                    }
-                }
-            }
-        }
-        for id in stoppedIDs { setTransitioning(id, false) }
-        await loadContainersFromDocker(docker: docker)
-    }
-
-    func stopContainersDocker(_ ids: [String], docker: DockerClient?) async {
-        guard let docker else { return }
-        let runningIDs = ids.filter { id in
-            containers.first(where: { $0.id == id })?.isRunning == true
-        }
-        for id in runningIDs { setTransitioning(id, true) }
-        await withTaskGroup(of: Void.self) { group in
-            for id in runningIDs {
-                group.addTask { [weak self] in
-                    do {
-                        _ = try await docker.api.ContainerStop(path: .init(id: id))
-                        await self?.setContainerRunningState(id, isRunning: false)
-                    } catch {
-                        Log.container.error(
-                            "Error stopping container \(id, privacy: .private): \(error.localizedDescription, privacy: .private)"
-                        )
-                        ErrorReporting.capture(error, domain: .container, operation: "batch_stop")
-                    }
-                }
-            }
-        }
-        for id in runningIDs { setTransitioning(id, false) }
-        await loadContainersFromDocker(docker: docker)
-    }
-
-    func removeContainersDocker(_ ids: [String], docker: DockerClient?) async {
-        guard let docker else { return }
-        for id in ids { setTransitioning(id, true) }
-        await withTaskGroup(of: Void.self) { group in
-            for id in ids {
-                group.addTask { [weak self] in
-                    do {
-                        _ = try await docker.api.ContainerDelete(path: .init(id: id), query: .init(force: true))
-                        await self?.removeContainerLocally(id)
-                    } catch {
-                        Log.container.error(
-                            "Error removing container \(id, privacy: .private): \(error.localizedDescription, privacy: .private)"
-                        )
-                        ErrorReporting.capture(error, domain: .container, operation: "batch_remove")
-                    }
-                }
-            }
-        }
-        NotificationCenter.default.post(name: .dockerDataChanged, object: nil)
-        await loadContainersFromDocker(docker: docker)
-    }
-
-    fileprivate static func normalized(_ value: String?) -> String? {
-        guard let trimmed = value?.trimmingCharacters(in: .whitespacesAndNewlines), !trimmed.isEmpty else {
-            return nil
-        }
-        return trimmed
-    }
-}
-
-// MARK: - Proto → UI Model Conversion
-
-extension ContainerViewModel {
-    /// Create a ContainerViewModel from a gRPC ContainerSummary.
-    init(from summary: Arcbox_V1_ContainerSummary) {
-        let name =
-            summary.names.first.map {
-                $0.hasPrefix("/") ? String($0.dropFirst()) : $0
-            } ?? summary.id.prefix(12).description
-
-        let state: ContainerState =
-            switch summary.state {
-            case "running": .running
-            case "paused": .paused
-            case "restarting": .restarting
-            case "dead": .dead
-            default: .stopped
-            }
-
-        let ports = summary.ports.map { port in
-            PortMapping(
-                hostPort: UInt16(port.hostPort),
-                containerPort: UInt16(port.containerPort),
-                protocol: port.protocol
-            )
-        }
-
-        let composeProject = summary.labels["com.docker.compose.project"]
-        let composeService = summary.labels["com.docker.compose.service"]
-        let rootfsMountPath = ContainerViewModel.inferRootFSMountPath(
-            explicitPath: nil,
-            labels: summary.labels,
-            mounts: []
-        )
-
-        self.init(
-            id: summary.id,
-            name: name,
-            image: summary.image,
-            state: state,
-            ports: ports,
-            createdAt: Date(timeIntervalSince1970: TimeInterval(summary.created)),
-            composeProject: composeProject,
-            composeService: composeService,
-            labels: summary.labels,
-            cpuPercent: 0,
-            memoryMB: 0,
-            memoryLimitMB: 0,
-            rootfsMountPath: rootfsMountPath
-        )
-    }
-
-    /// Create a ContainerViewModel from a Docker Engine API ContainerSummary.
-    init(fromDocker summary: Components.Schemas.ContainerSummary) {
-        let name =
-            summary.Names?.first.map {
-                $0.hasPrefix("/") ? String($0.dropFirst()) : $0
-            } ?? summary.Id?.prefix(12).description ?? "unknown"
-
-        let state: ContainerState =
-            switch summary.State?.lowercased() {
-            case "running": .running
-            case "paused": .paused
-            case "restarting": .restarting
-            case "dead": .dead
-            default: .stopped  // created, exited, removing -> stopped
-            }
-
-        let ports = (summary.Ports ?? []).compactMap { port -> PortMapping? in
-            guard let publicPort = port.PublicPort else { return nil }
-            return PortMapping(
-                hostPort: UInt16(publicPort),
-                containerPort: UInt16(port.PrivatePort),
-                protocol: port._Type.rawValue
-            )
-        }
-
-        let labels = summary.Labels?.additionalProperties ?? [:]
-        let composeProject = labels["com.docker.compose.project"]
-        let composeService = labels["com.docker.compose.service"]
-        let mounts = (summary.Mounts ?? []).compactMap { mount -> ContainerMount? in
-            guard let destination = ContainersViewModel.normalized(mount.Destination) else { return nil }
-            let source = ContainersViewModel.normalized(mount.Source) ?? "-"
-            return ContainerMount(
-                type: "unknown",
-                source: source,
-                destination: destination,
-                isReadOnly: false
-            )
-        }
-        let rootfsMountPath = ContainerViewModel.inferRootFSMountPath(
-            explicitPath: nil,
-            labels: labels,
-            mounts: mounts
-        )
-
-        self.init(
-            id: summary.Id ?? "",
-            name: name,
-            image: summary.Image ?? "",
-            state: state,
-            ports: ports,
-            createdAt: Date(timeIntervalSince1970: TimeInterval(summary.Created ?? 0)),
-            composeProject: composeProject,
-            composeService: composeService,
-            labels: labels,
-            cpuPercent: 0,
-            memoryMB: 0,
-            memoryLimitMB: 0,
-            mounts: mounts,
-            rootfsMountPath: rootfsMountPath
-        )
-    }
-}
